test(zqlite): isolate query scenario fixtures
